Moved clean function to views to keep functionality bundled Updated author and copyright information Autoformatted into pep8
